+++ tests/unit/slackNotification/test_slack_notification.py
@@ -1,8 +1,8 @@
 import os
 import sys
 import unittest
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../src')))
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../../src/common/python')))
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__),'../../../src')))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__),'../../../src/common/python')))
 from mock import patch, Mock
 from requests.exceptions import HTTPError
 from slackNotification import slack_notification
